backend/requirements.txt
# Web framework
Flask==3.0.3
flask-cors==4.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9

# Validation
pydantic==2.7.1

# HTTP client (for HuggingFace API calls)
requests==2.32.2

# Testing
pytest==8.2.0
pytest-flask==1.3.0

# Environment variables
python-dotenv==1.0.1

huggingface_hub==0.23.4
